+++ pom.xml
@@ -12,7 +12,7 @@
 	<properties>
 		<maven.compiler.target>1.8</maven.compiler.target>
 		<maven.compiler.source>1.8</maven.compiler.source>
-		<sonar.projectKey>csys-fresher-batch-2021_your_project_name</sonar.projectKey>
+		<sonar.projectKey>csys-fresher-batch-2021_primeapp-core-venkat</sonar.projectKey>
 		<sonar.organization>csys-fresher-batch-2021</sonar.organization>
 		<sonar.host.url>https://sonarcloud.io</sonar.host.url>
 	</properties>
