+++ src/main/java/in/venkat/exceptions/EmptyFieldException.java
@@ -1,6 +1,7 @@
 package in.venkat.exceptions;
 
 public class EmptyFieldException extends Exception {
+
 	private static final long serialVersionUID = 1L;
 
 	public EmptyFieldException(String message) {
