+++ src/main/java/in/venkat/exceptions/InvalidDetailsException.java
@@ -1,6 +1,7 @@
 package in.venkat.exceptions;
 
 public class InvalidDetailsException extends Exception {
+
 	private static final long serialVersionUID = 1L;
 
 	public InvalidDetailsException(String message) {
