+++ src/main/java/in/venkat/exceptions/InvalidMovieIdException.java
@@ -1,6 +1,7 @@
 package in.venkat.exceptions;
 
 public class InvalidMovieIdException extends Exception {
+
 	private static final long serialVersionUID = 1L;
 
 	public InvalidMovieIdException(String message) {
