+++ src/main/java/in/venkat/exceptions/InvalidNameException.java
@@ -1,6 +1,7 @@
 package in.venkat.exceptions;
 
 public class InvalidNameException extends Exception {
+
 	private static final long serialVersionUID = 1L;
 
 	public InvalidNameException(String message) {
