+++ src/main/java/in/venkat/util/ConnectionUtil.java
@@ -12,15 +12,19 @@ public class ConnectionUtil {
 	private static final String DB_PASSWORD = "19111999";
 	private static final String HOST = "localhost";
 	private static final int PORT = 5432;
-	private static final String DB_URL = "jdbc:postgresql://" + HOST + ":" + PORT + "/" + DATABASE_NAME; // jdbc:postgres://localhost:5432/shoppingapp_db
+	private static final String DB_URL = "jdbc:postgresql://" + HOST + ":" + PORT + "/" + DATABASE_NAME;
 
 	public static Connection getConnection() {
 
 		Connection connection = null;
 		try {
-			// Step 1: Load the database driver into memory ( ClassNotFoundException )
+			/**
+			 * Step 1: Load the database driver into memory ( ClassNotFoundException )
+			 */
 			Class.forName(DRIVER_CLASS_NAME);
-			// Step 2: Get the Database Connection (SQLException)
+			/**
+			 * Step 2: Get the Database Connection (SQLException)
+			 */
 			connection = DriverManager.getConnection(DB_URL, DB_USERNAME, DB_PASSWORD);
 			System.out.println(connection);
 		} catch (ClassNotFoundException | SQLException e) {
@@ -31,7 +35,9 @@ public class ConnectionUtil {
 	}
 
 	public static void close(PreparedStatement pst, Connection con) throws Exception {
-		// Null Check - to avoid Null Pointer Exception
+		/**
+		 * Null Check - to avoid Null Pointer Exception
+		 */
 		if (pst != null) {
 			pst.close();
 		}
