+++ src/test/java/in/venkat/searchServiceTest/ConnectionUtilTest.java
@@ -1,0 +1,16 @@
+package in.venkat.searchServiceTest;
+
+import static org.junit.Assert.fail;
+
+import in.venkat.util.ConnectionUtil;
+
+public class ConnectionUtilTest {
+
+	public static void main(String[] args) {
+		try {
+			ConnectionUtil.getConnection();
+		} catch (Exception e) {
+			fail();
+		}
+	}
+}
