--- src/test/java/in/venkat/test/ConnectionUtilTest.java
@@ -1,16 +1,0 @@
-package in.venkat.test;
-
-import static org.junit.Assert.fail;
-
-import in.venkat.util.ConnectionUtil;
-
-public class ConnectionUtilTest {
-
-	public static void main(String[] args) {
-		try {
-			ConnectionUtil.getConnection();
-		} catch (Exception e) {
-			fail();
-		}
-	}
-}
